+++ CHANGELOG.md
@@ -37,11 +37,16 @@ continue to use `react-loading-skeleton` v2.
 -   Add `enableAnimation` prop to allow disabling the animation
 -   Add `containerClassName` prop to allow customizing the container element
 -   Add `containerTestId` to make testing easier
+-   Add `aria-live` and `aria-busy` attributes to the skeleton container to
+    improve screen reader support
+
+### Other Changes
+
 -   Change the default `duration` from 1.2 s to 1.5 s
 -   Make the default `Skeleton` base color a _tiny_ bit darker so that the
     animation is more visible
--   Add `aria-live` and `aria-busy` attributes to the skeleton container to
-    improve screen reader support
+-   No longer require `width` and `height` to be set for the `circle` prop to
+    work
 
 ### Bug Fixes